prepare for tests
